+++ test/data/simple.js
@@ -4,6 +4,11 @@
  * @module simple
  */
 
+
+const INTERNAL_TWELVE = 12;
+// WARNING: Don't insert something between the module comment and the above constant. This is a regression
+// test for https://github.com/LaxarJS/laxar-dox/issues/21
+
 /**
  * The default for defaults.
  *
@@ -24,5 +29,5 @@ export const DEFAULT = 'default';
  *    a mostly constant `12`
  */
 export function func( a, b ) {
-   return 12;
+   return INTERNAL_TWELVE;
 }